feat: Fixing Shadow Acne

+++ Rust/src/algorithm/hittable.rs
@@ -6,12 +6,12 @@ use super::vec3::Vec3;
 pub struct HitRecord {
   pub point: Vec3,
   pub normal: Vec3,
-  pub t: f32,
+  pub t: f64,
   pub front_face: bool,
 }
 
 impl HitRecord {
-  pub fn new(point: Vec3, t: f32, normal: Vec3, ray: Ray) -> HitRecord {
+  pub fn new(point: Vec3, t: f64, normal: Vec3, ray: Ray) -> HitRecord {
     let front_face = ray.direction.dot(normal) < 0.;
     let normal = if front_face { normal } else { -normal };
     HitRecord {
