+++ Rust/src/algorithm/interval.rs
@@ -1,10 +1,10 @@
 pub struct Interval {
-  pub min: f32,
-  pub max: f32,
+  pub min: f64,
+  pub max: f64,
 }
 
 impl Interval {
-  pub fn new(min: f32, max: f32) -> Interval {
+  pub fn new(min: f64, max: f64) -> Interval {
     Interval { min, max }
   }
 
@@ -14,24 +14,24 @@ impl Interval {
 
   pub fn universe() -> Interval {
     Interval {
-      min: -f32::INFINITY,
-      max: f32::INFINITY,
+      min: -f64::INFINITY,
+      max: f64::INFINITY,
     }
   }
 
-  pub fn size(&self) -> f32 {
+  pub fn size(&self) -> f64 {
     self.max - self.min
   }
 
-  pub fn contains(&self, x: f32) -> bool {
+  pub fn contains(&self, x: f64) -> bool {
     self.min <= x && x <= self.max
   }
 
-  pub fn surrounds(&self, x: f32) -> bool {
+  pub fn surrounds(&self, x: f64) -> bool {
     self.min < x && x < self.max
   }
 
-  pub fn clamp(&self, x: f32) -> f32 {
+  pub fn clamp(&self, x: f64) -> f64 {
     if x < self.min {
       return self.min;
     }
@@ -45,8 +45,8 @@ impl Interval {
 impl Default for Interval {
   fn default() -> Self {
     Interval {
-      min: f32::INFINITY,
-      max: -f32::INFINITY,
+      min: f64::INFINITY,
+      max: -f64::INFINITY,
     }
   }
 }
